refactor attachment.ts
